backend/requirements.txt
## `backend/requirements.txt`

```txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.4
pydantic-settings==2.3.4
sqlalchemy==2.0.30
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
httpx==0.27.0
anyio==4.4.0

# Optional provider SDKs (installed but only used if API keys set)
openai==1.30.5
anthropic==0.30.0
google-generativeai==0.7.2

# utilities
tenacity==8.2.3
```

---